Extract social media buttons to separate component
