+++ lib/ui/views/login/login_view.dart
@@ -1,14 +1,12 @@
-import 'dart:io';
-
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:nowu/assets/constants.dart';
 import 'package:nowu/themes.dart';
 import 'package:nowu/ui/common/form.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked/stacked_annotations.dart';
 
+import 'components/social_media_buttons.dart';
 import 'login_view.form.dart';
 import 'login_viewmodel.dart';
 
@@ -115,24 +113,7 @@ class LoginView extends StackedView<LoginViewModel> with $LoginView {
                   ),
                 ),
                 const SizedBox(height: 30),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceAround,
-                  children: [
-                    Platform.isIOS
-                        ? SocialMediaLoginButton(
-                            onPressed: model.loginWithApple,
-                            iconData: FontAwesomeIcons.apple,
-                          )
-                        : SocialMediaLoginButton(
-                            onPressed: model.loginWithGoogle,
-                            iconData: FontAwesomeIcons.google,
-                          ),
-                    SocialMediaLoginButton(
-                      onPressed: model.loginWithFacebook,
-                      iconData: FontAwesomeIcons.facebookF,
-                    ),
-                  ],
-                ),
+                SocialMediaLoginButtons(model),
                 const SizedBox(height: 30),
                 Container(
                   width: MediaQuery.of(context).size.width * 0.9,
@@ -200,31 +181,6 @@ class LoginView extends StackedView<LoginViewModel> with $LoginView {
               _loginForm(viewModel),
             ],
           ),
-        ),
-      ),
-    );
-  }
-}
-
-class SocialMediaLoginButton extends StatelessWidget {
-  final VoidCallback onPressed;
-  final IconData iconData;
-
-  const SocialMediaLoginButton({
-    Key? key,
-    required this.onPressed,
-    required this.iconData,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return TextButton(
-      onPressed: onPressed,
-      child: Icon(iconData, color: Colors.orange),
-      style: ButtonStyle(
-        backgroundColor: MaterialStateProperty.all(Colors.white),
-        padding: const MaterialStatePropertyAll(
-          EdgeInsets.symmetric(vertical: 15, horizontal: 60),
         ),
       ),
     );
